TESTS FAIL
Revert "Changed the expectation that every technique.Find() will close its outbound resultsChan when done. Required changing a few expectations throughout. This should allow us to fix the last bug preventing the switch from NewSearch --> Search"

We'll be moving to a model where we signal the change a different way

This reverts commit 8e38905eb30b0ca79b5d7d6c0b14c64a45bd821d.